+++ src/components/Logo/Logo.css
@@ -1,0 +1,17 @@
+.logo {
+  cursor: pointer;
+  transition: all 0.4s;
+}
+
+@media screen and (max-width: 1280px) {
+  .logo {
+    left: 30px;
+    z-index: 0;
+  }
+}
+
+@media screen and (max-width: 600px) {
+  .logo {
+    left: 14px;
+  }
+}
