+++ src/index.css
@@ -1,7 +1,10 @@
 @import url(./vendor/normalize.css);
 @import url(./vendor/fonts/inter.css);
+/* - */
 @import url(./components/App/App.css);
 @import url(./components/Header/Header.css);
+@import url(./components/Logo/Logo.css);
+@import url(./components/Navigation/Navigation.css);
 @import url(./components/Main/Main.css);
 @import url(./components/Footer/Footer.css);
 @import url(./components/NotFound/NotFound.css);
@@ -20,3 +23,4 @@
 /* фильмы */
 @import url(./components/Profile/Profile.css);
 @import url(./components/UserForm/UserForm.css);
+/* сервис */
